refactor: consolidate PatchableVulnerability type
Create src/lib/patcher/types.ts with shared PatchableVulnerability
interface. Update 3 patcher files to import from shared location.
