+++ src/app/api/docker/info/route.ts
@@ -1,5 +1,6 @@
 import { NextResponse } from 'next/server';
 import { checkDockerAccess, getSwarmInfo } from '@/lib/docker';
+import { apiError } from '@/lib/api-utils';
 
 export async function GET() {
   try {
