+++ src/components/scan/VulnerabilityDetailModal.tsx
@@ -23,8 +23,8 @@ import {
   IconFingerprint,
   IconLink,
 } from "@tabler/icons-react";
-import { toast } from "sonner";
 import { formatDate, renderValue } from "@/lib/format-utils";
+import { copyToClipboard } from "@/lib/clipboard";
 
 interface VulnerabilityDetailModalProps {
   isOpen: boolean;
@@ -40,11 +40,6 @@ export function VulnerabilityDetailModal({
   classification,
 }: VulnerabilityDetailModalProps) {
   if (!vulnerability) return null;
-
-  const copyToClipboard = (text: string, label: string) => {
-    navigator.clipboard.writeText(text);
-    toast.success(`${label} copied to clipboard`);
-  };
 
   const getSeverityColor = (severity: string) => {
     const colors: Record<string, string> = {
