+++ src/components/vulnerability-details-modal.tsx
@@ -26,8 +26,8 @@ import {
   IconCheck,
   IconCopy,
 } from "@tabler/icons-react";
-import { toast } from "sonner";
 import { getSeverityBadgeVariant } from "@/lib/severity-utils";
+import { copyToClipboard } from "@/lib/clipboard";
 
 interface VulnerabilityDetailsModalProps {
   vulnerability: {
@@ -61,11 +61,6 @@ export function VulnerabilityDetailsModal({
   isOpen,
   onClose,
 }: VulnerabilityDetailsModalProps) {
-  const copyToClipboard = (text: string) => {
-    navigator.clipboard.writeText(text);
-    toast.success("Copied to clipboard");
-  };
-
   if (!vulnerability) return null;
 
   // Separate truly affected images from false positives
