+++ src/lib/patcher/PatchExecutorTar.ts
@@ -8,14 +8,17 @@ import { PatchStrategy } from './strategies/PatchStrategy';
 import { AptPatchStrategy } from './strategies/AptPatchStrategy';
 import { YumPatchStrategy } from './strategies/YumPatchStrategy';
 import { ApkPatchStrategy } from './strategies/ApkPatchStrategy';
-import type { 
-  PatchOperation, 
-  PatchOperationStatus, 
+import type {
+  PatchOperation,
+  PatchOperationStatus,
   PatchResult,
-  PatchStrategy as PatchStrategyType 
+  PatchStrategy as PatchStrategyType
 } from '@/generated/prisma';
+import type { PatchableVulnerability } from './types';
 
 const execAsync = promisify(exec);
+
+export type { PatchableVulnerability };
 
 export interface PatchRequest {
   sourceImageId: string;
@@ -24,14 +27,6 @@ export interface PatchRequest {
   targetTag?: string;
   dryRun?: boolean;
   selectedVulnerabilityIds?: string[];
-}
-
-export interface PatchableVulnerability {
-  cveId: string;
-  packageName: string;
-  currentVersion: string;
-  fixedVersion: string;
-  packageManager: string;
 }
 
 export class PatchExecutorTar {
