+++ src/lib/patcher/types.ts
@@ -1,0 +1,8 @@
+export interface PatchableVulnerability {
+  id?: string;
+  cveId: string;
+  packageName: string;
+  currentVersion: string;
+  fixedVersion: string;
+  packageManager: string;
+}
